LLM_Project_OnHand/Bili-Agent/Multi-Agent/requirements.txt
langchain~=0.3.3
langchain_community
langchain_openai
faiss-cpu
langgraph~=0.2.35
fastapi~=0.115.2
sse_starlette
streamlit~=1.39.0
bilibili-api-dev
langserve~=0.3.0
zhipuai~=2.1.5.20241204
httpx-sse
pillow~=10.1.0
python-dotenv~=1.0.1
uvicorn~=0.32.1
pydantic~=2.10.3
openai~=1.57.0
aiohttp~=3.9.5
tqdm~=4.66.6
colorama~=0.4.6
httpx~=0.25.2
yarl~=1.9.11
brotli~=1.1.0
pyyaml~=6.0.2
rsa~=4.9
qrcode~=7.4.2
beautifulsoup4~=4.12.3
apscheduler~=3.10.4